FIX: Amend failing reviewable system test selector (#120)
This system test is failing in core after some changes to the review queue.

This PR fixes that by relying on the text of the button instead of a class name. This should be backwards compatible since the content hasn't change, and arguably a little bit more clear and a little bit more stable than relying on the class name.